requirements.txt
biopython==1.78
gensim==3.8.3
numpy==1.20.1
scipy==1.6.1
six==1.15.0
smart-open==4.2.0
tqdm==4.57.0
